chore(assets): rename icons folder to assets
