+++ src/components/assets/DeleteIcon.svelte
