+++ src/components/assets/MoonIcon.svelte
