+++ src/components/assets/SunIcon.svelte
